commit_second_checkup
